Ensure that items index only finds items owned by the user
Specify inside the Item.find method on the items controller index
that it is looking for items with the '_owner' property that matches
the user ID of the person making the request.

Accordingly remove the index exception from the authentication
method inside the controller object export.

@quidprocrow
@briankeegan

(Note that Brian has been present for the past two commits, I simply
forgot.)