Ensure that the items show method works for authenticated users only
Ensure that items show method works for authenticated users by including
show in the before action where the model and the `{ forUser: true }` are
set, such that only items that belong to that user can be shown to user,
and no other users, authenticated or otherwise, can view it.

In honor of @smb2255, this has been a long day of commits, and this
particular problem took us longer than it should have, because of
the 'black box' quality of the before methods in the controller object
export. @briankeegan's heart almost broke when I proposed achieving
essentially the same set of data by making a find request, like index,
but with the object's id set to the request's id. That was, admittedly,
inelegant but would have worked. As it is, the code is minimal,
and the day is done, and the night is dark, and it is time
to
call
it a
night.

@quidprocrow
@briankeegan